procs: return a slice of bools
When we process the procs, we now use the withthreads
string to see if we should also gather thread stats for that
pid. function now returns two slices - one containing the
procs and the other containing a bool for the associated proc to determine whether the thread level stats are
required.

Signed-off-by: Paul Cuzner <pcuzner@ibm.com>